Extract some OS-specific socket details into common_socket_impl
