+++ task1-2/Makefile
@@ -3,7 +3,7 @@
 # Based on https://github.com/yeputons/project-templates
 
 TARGETS=bin/test32 bin/test64 bin/client32 bin/client64 bin/server32 bin/server
-SRCS_common=$(SRCDIR)/tcp_socket.cpp $(SRCDIR)/protocol.cpp
+SRCS_common=$(SRCDIR)/tcp_socket.cpp $(SRCDIR)/common_socket_impl.cpp $(SRCDIR)/protocol.cpp
 SRCS_test=$(SRCS_common) $(SRCDIR)/test.cpp $(SRCDIR)/test_protocol.cpp
 SRCS_client=$(SRCS_common) $(SRCDIR)/client.cpp
 SRCS_server=$(SRCS_common) $(SRCDIR)/server.cpp
